ticket_manager/pyproject.toml
[project]
name = "ticket-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "ghaluccas",email = "gabsgit@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.9.3"
sqlalchemy = "^2.0.37"
uvicorn = "^0.34.0"
alembic = "^1.14.1"
httpx = "^0.28.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations' , 'alembic']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix"
run = "uvicorn ticket_manager.app:app --reload"
pre_test = "task lint"
test = "pytest -s -x --cov=ticket_manager -vv"
post_test = "coverage html"


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

